backend/requirements.txt
fastapi
uvicorn
supervision
python-multipart
opencv-python-headless
numpy==1.26.*
python-dotenv
pandas
yt-dlp
google-api-python-client
huggingface-hub
iopath
Pillow
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
